Update main.dart : Cleaned UP & Moved Main Page to seperated class
